Move common type declaration into its own file.
