+++ assets/js/modules/analytics-4/components/site-goals/notifications/IntroModalBanner/IntroModalLead.tsx
@@ -29,7 +29,7 @@ import BannerModal from '@/js/components/BannerModal/index';
 // @ts-expect-error - We need to add types for imported SVGs.
 import SiteGoalsIntroModalLeadGraphic from '@/svg/graphics/site-goals-intro-modal-lead.svg';
 import Link from '@/js/components/Link';
-import { IntroModalVariantProps } from '@/js/modules/analytics-4/components/site-goals/notifications/IntroModalBanner/index';
+import type { IntroModalVariantProps } from './types';
 
 export default function IntroModalLead( {
 	onView,
