+++ assets/js/modules/analytics-4/components/site-goals/notifications/IntroModalBanner/types.ts
@@ -1,0 +1,24 @@
+/**
+ * IntroModalBanner types.
+ *
+ * Site Kit by Google, Copyright 2026 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+export interface IntroModalVariantProps {
+	onView: () => void;
+	onConfirm: () => void;
+	onClickLearnMore: () => void;
+	onDismiss: () => void;
+}
